・srcフォルダ下のcomponentsフォルダへ移行 ・cssの形式をTailwindへ変更
